+++ src/main/java/com/flab/CafeMap/domain/user/service/UserAddressService.java
@@ -6,8 +6,8 @@ import com.flab.CafeMap.domain.user.dao.UserAddressMapper;
 import com.flab.CafeMap.domain.user.dao.UserMapper;
 import com.flab.CafeMap.web.api.KakaoMapApi;
 import com.flab.CafeMap.web.user.dto.UserAddressSaveRequest;
-import com.flab.CafeMap.web.api.dto.KakaoMapApiRequest;
-import com.flab.CafeMap.web.api.dto.KakaoMapApiResponse;
+import com.flab.CafeMap.web.user.dto.KakaoMapApi.KakaoMapApiRequest;
+import com.flab.CafeMap.web.user.dto.KakaoMapApi.KakaoMapApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
