+++ src/main/java/com/flab/CafeMap/web/user/dto/KakaoMapApi/Meta.java
@@ -1,4 +1,4 @@
-package com.flab.CafeMap.web.api.dto;
+package com.flab.CafeMap.web.user.dto.KakaoMapApi;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
